moved export away from JSX to allow HMR update (on the fly) added dev hmr for player only gui
