+++ react-client/package.json
@@ -5,10 +5,11 @@
   "proxy": "http://localhost:9001",
   "type": "module",
   "scripts": {
-    "dev": "vite build --watch",
+    "dev": "vite",
     "build": "tsc && eslint . && vite build",
     "lint": "eslint . --max-warnings 0",
     "preview": "vite preview",
+    "dev:player": "vite --config vite.config.player.ts",
     "test:runserver:macos": "export RUNNING_TESTS=true; npm run private:removeconfig:macos;cd ..;./runtestserver.sh",
     "test:prepareserver:macos": "cd ../;mvn -P macos package -DskipTests -Doffline -Djaronly;cd react-client;npm run private:copy-jar;npm run private:ensure-file-permissions",
     "test:prepareserver:macos-arm": "cd ../;mvn -P macos-arm package -DskipTests -Doffline;cd react-client;npm run private:copy-jar;npm run private:ensure-file-permissions",
@@ -48,7 +49,6 @@
     "eslint-plugin-react-refresh": "0.4.19",
     "typescript": "5.7.3",
     "typescript-eslint": "8.25.0",
-    "vite": "6.2.0",
-    "vite-live-preview": "0.3.2"
+    "vite": "6.2.0"
   }
 }
