+++ react-client/src/UmsApp.tsx
@@ -40,12 +40,12 @@ import SharedContent from './components/SharedContent/SharedContent';
 import UserMenu from './components/UserMenu/UserMenu';
 import MainContext from './contexts/main-context';
 import SessionContext from './contexts/session-context';
-import { AccountsProvider } from './providers/accounts-provider';
-import { I18nProvider } from './providers/i18n-provider';
-import { MainProvider } from './providers/main-provider';
-import { PlayerEventProvider } from './providers/player-server-event-provider';
-import { ServerEventProvider } from './providers/server-event-provider';
-import { SessionProvider } from './providers/session-provider';
+import AccountsProvider from './providers/accounts-provider';
+import I18nProvider from './providers/i18n-provider';
+import MainProvider from './providers/main-provider';
+import PlayerEventProvider from './providers/player-server-event-provider';
+import ServerEventProvider from './providers/server-event-provider';
+import SessionProvider from './providers/session-provider';
 import { havePermission, Permissions } from './services/accounts-service';
 import { refreshAuthTokenNearExpiry, setAxiosAuthorization } from './services/auth-service';
 
@@ -65,81 +65,79 @@ function UmsApp() {
           {main => (
             <SessionProvider><SessionContext.Consumer>
               {session => (
-                <div dir={dir} className='bodyBackgroundImageScreen'>
-                  <AppShell
-                    padding='md'
-                    navbar={main.navbarValue ? {
-                      width: { sm: 200, lg: 300 },
-                      breakpoint: 'sm',
-                      collapsed: { mobile: !main.navbarOpened, desktop: false }
-                    } : undefined}
-                    header={{ height: 50 }}
-                    styles={(theme) => ({
-                      main: { backgroundColor: computedColorScheme === 'dark' ? theme.colors.darkTransparent[8] : theme.colors.lightTransparent[0] },
-                    })}
-                  >
-                    {main.navbarValue && <AppShell.Navbar
-                      p='xs'
-                      style={(theme: MantineTheme) => ({ backgroundColor: computedColorScheme === 'dark' ? theme.colors.darkTransparent[8] : theme.colors.lightTransparent[0] })}
+                <Router>
+                  <div dir={dir} className='bodyBackgroundImageScreen'>
+                    <AppShell
+                      padding='md'
+                      navbar={main.navbarValue ? {
+                        width: { sm: 200, lg: 300 },
+                        breakpoint: 'sm',
+                        collapsed: { mobile: !main.navbarOpened, desktop: false }
+                      } : undefined}
+                      header={{ height: 50 }}
+                      styles={(theme) => ({
+                        main: { backgroundColor: computedColorScheme === 'dark' ? theme.colors.darkTransparent[8] : theme.colors.lightTransparent[0] },
+                      })}
                     >
-                      <AppShell.Section grow my="md" component={ScrollArea}>
-                        <Stack gap={0}>{main.navbarValue}</Stack>
-                      </AppShell.Section>
-                    </AppShell.Navbar>}
-                    <AppShell.Header
-                      p='xs'
-                      style={(theme) => ({ backgroundColor: computedColorScheme === 'dark' ? theme.colors.darkTransparent[8] : theme.colors.lightTransparent[0] })}
-                    >
-                      <Group justify='space-between'>
-                        <Group justify='left'>
-                          {main.navbarValue &&
-                            <Burger
-                              hiddenFrom='sm'
-                              opened={main.navbarOpened}
-                              onClick={() => main.setNavbarOpened((o: boolean) => !o)}
-                              size='sm'
-                              mr='xl'
-                            />
-                          }
-                          {session.account && session.account.user &&
-                            <Button
-                              variant='transparent'
-                              style={() => ({ cursor: 'default', color: computedColorScheme === 'dark' ? 'white' : 'black' })}
-                              leftSection={
-                                <Avatar radius='sm' size='sm' src={session.account.user.avatar !== '' ? session.account.user.avatar : null}>
-                                  {session.account.user.avatar === '' && <IconUser size={16} />}
-                                </Avatar>
-                              }
-                            >
-                              {session.account.user.displayName}
-                            </Button>
-                          }
+                      {main.navbarValue && <AppShell.Navbar
+                        p='xs'
+                        style={(theme: MantineTheme) => ({ backgroundColor: computedColorScheme === 'dark' ? theme.colors.darkTransparent[8] : theme.colors.lightTransparent[0] })}
+                      >
+                        <AppShell.Section grow my="md" component={ScrollArea}>
+                          <Stack gap={0}>{main.navbarValue}</Stack>
+                        </AppShell.Section>
+                      </AppShell.Navbar>}
+                      <AppShell.Header
+                        p='xs'
+                        style={(theme) => ({ backgroundColor: computedColorScheme === 'dark' ? theme.colors.darkTransparent[8] : theme.colors.lightTransparent[0] })}
+                      >
+                        <Group justify='space-between'>
+                          <Group justify='left'>
+                            {main.navbarValue &&
+                              <Burger
+                                hiddenFrom='sm'
+                                opened={main.navbarOpened}
+                                onClick={() => main.setNavbarOpened((o: boolean) => !o)}
+                                size='sm'
+                                mr='xl'
+                              />
+                            }
+                            {session.account && session.account.user &&
+                              <Button
+                                variant='transparent'
+                                style={() => ({ cursor: 'default', color: computedColorScheme === 'dark' ? 'white' : 'black' })}
+                                leftSection={
+                                  <Avatar radius='sm' size='sm' src={session.account.user.avatar !== '' ? session.account.user.avatar : null}>
+                                    {session.account.user.avatar === '' && <IconUser size={16} />}
+                                  </Avatar>
+                                }
+                              >
+                                {session.account.user.displayName}
+                              </Button>
+                            }
+                          </Group>
+                          <Group justify='right'>
+                            <ActionIcon variant='default' onClick={() => toggleColorScheme()} size={30}>
+                              {computedColorScheme === 'dark' ? <IconSun size={16} /> : <IconMoonStars size={16} />}
+                            </ActionIcon>
+                            <LanguagesMenu />
+                            {session.account && <UserMenu />}
+                          </Group>
                         </Group>
-                        <Group justify='right'>
-                          <ActionIcon variant='default' onClick={() => toggleColorScheme()} size={30}>
-                            {computedColorScheme === 'dark' ? <IconSun size={16} /> : <IconMoonStars size={16} />}
-                          </ActionIcon>
-                          <LanguagesMenu />
-                          {session.account && <UserMenu />}
-                        </Group>
-                      </Group>
-                    </AppShell.Header>
-                    <AppShell.Main>
-                      {(session.initialized && session.player) ? (
-                        (!session.authenticate || session.account) ? (
-                          <Router>
+                      </AppShell.Header>
+                      <AppShell.Main>
+                        {(session.initialized && session.player) ? (
+                          (!session.authenticate || session.account) ? (
                             <Routes>
                               <Route path='about' element={<About />}></Route>
                               <Route path='player' element={<PlayerEventProvider><Player /></PlayerEventProvider>}></Route>
                               <Route path='player/:req/:id' element={<PlayerEventProvider><Player /></PlayerEventProvider>}></Route>
                               <Route index element={<PlayerEventProvider><Player /></PlayerEventProvider>} />
                             </Routes>
-                          </Router>
-                        ) : (
-                          <PlayerLogin />
-                        )
-                      ) : session.account ? (
-                        <Router>
+                          ) : (
+                            <PlayerLogin />
+                          )
+                        ) : session.account ? (
                           <Routes>
                             <Route path='about' element={<ServerEventProvider><About /></ServerEventProvider>}></Route>
                             <Route path='accounts' element={<ServerEventProvider><AccountsProvider><Accounts /></AccountsProvider></ServerEventProvider>}></Route>
@@ -159,24 +157,24 @@ function UmsApp() {
                               element={<Navigate replace to='/' />}
                             />
                           </Routes>
-                        </Router>
-                      ) : session.initialized ? (
-                        <Login />
-                      ) : (
-                        <Center>
-                          <Box style={{ maxWidth: 1024 }} mx='auto'>
-                            <Loader size='xl' variant='dots' style={{ marginTop: '150px' }} />
-                          </Box>
-                        </Center>
+                        ) : session.initialized ? (
+                          <Login />
+                        ) : (
+                          <Center>
+                            <Box style={{ maxWidth: 1024 }} mx='auto'>
+                              <Loader size='xl' variant='dots' style={{ marginTop: '150px' }} />
+                            </Box>
+                          </Center>
+                        )}
+                      </AppShell.Main>
+                      {main.statusLine && (
+                        <AppShell.Footer>
+                          <Text>{main.statusLine}</Text>
+                        </AppShell.Footer>
                       )}
-                    </AppShell.Main>
-                    {main.statusLine && (
-                      <AppShell.Footer>
-                        <Text>{main.statusLine}</Text>
-                      </AppShell.Footer>
-                    )}
-                  </AppShell>
-                </div>
+                    </AppShell>
+                  </div>
+                </Router>
               )}
             </SessionContext.Consumer></SessionProvider>
           )}
