+++ react-client/src/components/Home/Home.tsx
@@ -28,6 +28,7 @@ import Renderers from './Renderers';
 import NetworkDevices from './NetworkDevices';
 import { renderersApiUrl } from '../../utils';
 import { havePermission, Permissions } from '../../services/accounts-service';
+import { NetworkDevicesFilter, Renderer, User } from '../../services/home-service';
 
 const Home = () => {
   const i18n = useContext(I18nContext);
@@ -257,53 +258,6 @@ const Home = () => {
 
 interface RendererAction extends Renderer {
   action: string,
-}
-
-export interface RendererState {
-  mute: boolean,
-  volume: number,
-  playback: number,
-  name: string,
-  uri: string,
-  metadata: string,
-  position: string,
-  duration: string,
-  buffer: number,
-}
-
-export interface Renderer {
-  id: number,
-  name: string,
-  address: string,
-  uuid: string,
-  icon: string,
-  playing: string,
-  time: string,
-  progressPercent: number,
-  isActive: boolean,
-  isAllowed: boolean,
-  isAuthenticated: boolean,
-  userId: number,
-  controls: number,
-  state: RendererState,
-}
-
-export interface NetworkDevice {
-  hostName: string,
-  ipAddress: string,
-  lastSeen: number
-}
-
-export interface NetworkDevicesFilter {
-  name: string,
-  isAllowed: boolean,
-  isDefault: boolean,
-  devices: NetworkDevice[]
-}
-
-export interface User {
-  value: number,
-  label: string
 }
 
 export default Home;
