+++ react-client/src/providers/main-provider.tsx
@@ -22,7 +22,7 @@ interface Props {
   children?: ReactNode
 }
 
-export const MainProvider = ({ children }: Props) => {
+const MainProvider = ({ children }: Props) => {
   const [navbarValue, setNavbarValue] = useState(undefined);
   const [navbarOpened, setNavbarOpened] = useState<boolean>(false);
   const [statusLine, setStatusLine] = useState(undefined);
@@ -38,3 +38,5 @@ export const MainProvider = ({ children }: Props) => {
     }}>{children}</Provider>
   )
 }
+
+export default MainProvider;
