+++ react-client/src/services/i18n-service.tsx
@@ -14,19 +14,23 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-import { Direction } from '@mantine/core';
-import { Context, createContext } from 'react';
+import { Direction } from "@mantine/core";
 
-import { I18nInterface, LanguageValue } from '../services/i18n-service';
+export interface I18nInterface {
+  get: (value: string) => string;
+  getI18nString: (value: string) => string;
+  getI18nFormat: (value: string[]) => string;
+  language: string;
+  dir: Direction;
+  languages: LanguageValue[];
+  setLanguage: (language: string) => void;
+}
 
-const I18nContext: Context<I18nInterface> = createContext({
-  get: (value: string) => { return value },
-  getI18nString: (value: string) => { return value },
-  getI18nFormat: (value: string[]) => { return value.length ? value[0] : '' },
-  language: 'en-US',
-  dir: 'ltr' as Direction,
-  languages: [] as LanguageValue[],
-  setLanguage: (_language: string) => { }
-});
+export interface LanguageValue {
+  id: string,
+  name: string,
+  defaultname: string,
+  country: string,
+  coverage: number,
+}
 
-export default I18nContext;
