+++ react-client/src/services/server-event-service.tsx
@@ -14,19 +14,26 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-import { Direction } from '@mantine/core';
-import { Context, createContext } from 'react';
+export interface UmsMemory {
+  max: number,
+  used: number,
+  dbcache: number,
+  buffer: number
+}
 
-import { I18nInterface, LanguageValue } from '../services/i18n-service';
+export interface ServerEventInterface {
+  connectionStatus: number;
+  memory: UmsMemory;
+  updateAccounts: boolean;
+  setUpdateAccounts: (updateAccounts: boolean) => void;
+  reloadable: boolean;
+  userConfiguration: any;
+  setUserConfiguration: (config: any) => void;
+  mediaScan: boolean;
+  hasRendererAction: boolean;
+  getRendererAction: () => any;
+  hasNewLogLine: boolean;
+  getNewLogLine: () => any;
+}
 
-const I18nContext: Context<I18nInterface> = createContext({
-  get: (value: string) => { return value },
-  getI18nString: (value: string) => { return value },
-  getI18nFormat: (value: string[]) => { return value.length ? value[0] : '' },
-  language: 'en-US',
-  dir: 'ltr' as Direction,
-  languages: [] as LanguageValue[],
-  setLanguage: (_language: string) => { }
-});
 
-export default I18nContext;
