+++ react-client/vite.config.ts
@@ -1,11 +1,9 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
-import livePreview from 'vite-live-preview'
 
 export default defineConfig({
   plugins: [
-    react(),
-    livePreview()
+    react()
   ],
   build: {
     outDir: '../src/main/external-resources/web/react-client',
@@ -32,6 +30,9 @@ export default defineConfig({
     open: '/',
     proxy: {
       '/v1': 'http://localhost:9001',
+    },
+    watch: {
+      usePolling: true
     }
   }
 })
